requirements.txt
langchain
langchain-community
langchain-text-splitters

langchain-google-genai
langchain-huggingface

chromadb
langchain-chroma

sentence-transformers
unstructured[pdf]

python-dotenv

